refactor(faz-15): split cli.py into cli/ package
Convert ``forgelm/cli.py`` (single ~2300-line file, 1.76× the
``architecture.md`` cohesion ceiling) into a ``forgelm/cli/`` package
following ``docs/analysis/code_reviews/split-design-data_audit-cli-202604300906.md``
§2 -- mirrors the Faz 14 split for ``data_audit.py``.

Layout (all modules well below the 1000-line ceiling):

- ``__init__.py``       facade re-exporting the full public + test-touched
                        surface (``# noqa: F401`` on private re-exports)
- ``__main__.py``       ``python -m forgelm.cli`` entry, load-bearing for
                        the quickstart subprocess flow
- ``_exit_codes.py``    EXIT_* + ``_PUBLIC_EXIT_CODES``
- ``_logging.py``       ``_setup_logging``, ``_get_version``, package logger
- ``_argparse_types.py``argparse type validators + shared subparser flags
- ``_resume.py``        ``_resolve_resume_checkpoint``
- ``_result.py``        ``_output_result`` + JSON envelope helpers
- ``_parser.py``        all 9 ``_add_*_subcommand`` registrars + ``parse_args``
- ``_dry_run.py``       ``_run_dry_run`` + dry-run-summary helpers
- ``_fit_check.py``     ``_run_fit_check``
- ``_no_train_modes.py``benchmark-only / merge / generate-data /
                        compliance-export + ``_maybe_run_no_train_mode``
- ``_wizard.py``        ``_maybe_run_wizard``
- ``_config_load.py``   ``_load_config_or_exit``, ``_apply_offline_flag``
- ``_training.py``      ``_run_training_pipeline`` + ``_report_training_error``
- ``_dispatch.py``      ``_dispatch_subcommand`` + ``main()``
- ``subcommands/``      one file per subcommand (chat / export / deploy /
                        ingest / audit / verify-audit / quickstart) + an
                        ``_approve.py`` co-locating the approve/reject
                        helper set since they share state.

Public API contract (split-design §2.b) preserved at
``forgelm.cli.<name>``:

  EXIT_SUCCESS, EXIT_CONFIG_ERROR, EXIT_TRAINING_ERROR,
  EXIT_EVAL_FAILURE, EXIT_AWAITING_APPROVAL,
  parse_args, main, _get_version, _resolve_resume_checkpoint,
  _run_dry_run, _setup_logging, _run_compliance_export, _run_fit_check,
  _run_data_audit, _output_result, _build_quickstart_inherited_flags

Test-touched private helpers also re-exported for ``patch(...)``
compatibility (``_run_chat_cmd``, ``_load_quickstart_train_paths``,
``_build_approval_notifier``, the four other approve/reject helpers,
the dispatchers themselves, the registrars, etc.).

To keep ``patch("forgelm.cli._run_*_cmd", ...)`` and
``patch("forgelm.cli._build_approval_notifier", ...)`` mocks effective
when called from inside the package, ``_dispatch_subcommand`` and the
approve / reject / audit / quickstart-chat dispatchers re-resolve their
helpers via ``from forgelm import cli as _cli_facade`` at call time.

Acceptance gate (all green):

- ``ruff format`` + ``ruff check forgelm/ tests/`` -- clean
- ``pytest tests/ -q`` -- 966 passed, 13 skipped (matches pre-split)
- ``forgelm --help`` -- byte-identical
- ``forgelm <sub> --help`` -- byte-identical for chat / audit /
  verify-audit / approve / reject / quickstart / export / deploy / ingest
- ``python3 -m forgelm.cli --help`` -- works (program name shifts to
  ``__main__.py`` per argparse default; functionally identical)
- ``forgelm --config config_template.yaml --dry-run`` -- exits 0
- ``forgelm --version`` -- unchanged
- Logger name preserved at ``forgelm.cli``

No test files needed import-path updates -- every existing
``from forgelm.cli import …`` and ``patch("forgelm.cli.…", …)``
resolves through the facade.
