+++ forgelm/cli/subcommands/__init__.py
@@ -1,0 +1,6 @@
+"""Subcommand dispatchers for the ForgeLM CLI.
+
+Each module here implements one ``forgelm <command>`` dispatcher. Heavy
+imports are deferred to the function bodies so ``forgelm --help`` and
+``--version`` stay lightweight.
+"""
