+++ forgelm/cli/subcommands/_export.py
@@ -1,0 +1,52 @@
+"""``forgelm export`` dispatcher."""
+
+from __future__ import annotations
+
+import json
+import sys
+
+from .._exit_codes import EXIT_TRAINING_ERROR
+from .._logging import logger
+
+
+def _run_export_cmd(args, output_format: str) -> None:
+    """Dispatch the ``forgelm export`` subcommand."""
+    from ...export import export_model
+
+    result = export_model(
+        model_path=args.model_path,
+        output_path=args.output,
+        output_format=args.format,
+        quant=args.quant,
+        adapter=args.adapter,
+        update_integrity=not args.no_integrity_update,
+    )
+
+    if output_format == "json":
+        print(
+            json.dumps(
+                {
+                    "success": result.success,
+                    "output_path": result.output_path,
+                    "format": result.format,
+                    "quant": result.quant,
+                    "sha256": result.sha256,
+                    "size_bytes": result.size_bytes,
+                    "error": result.error,
+                },
+                indent=2,
+            )
+        )
+    else:
+        if result.success:
+            logger.info(
+                "Export complete: %s (quant=%s, sha256=%s…)",
+                result.output_path,
+                result.quant,
+                (result.sha256 or "")[:12],
+            )
+        else:
+            logger.error("Export failed: %s", result.error)
+
+    if not result.success:
+        sys.exit(EXIT_TRAINING_ERROR)
